get_max_ploop_size(): change type of last arg
Use "unsigned long long" as a portable type, to fix the following
compilation error on POWER (caused by the fact that __u64 is defined
as "unsigned long" on POWER, but "unsigned long long" on x86_64).

> ploop.c: In function ‘ploop_get_max_size’:
> ploop.c:305:2: error: passing argument 3 of ‘get_max_ploop_size’ from incompatible pointer type [-Werror]
>   if (get_max_ploop_size(default_fmt_version(), blocksize, max))
>   ^
> ploop.c:275:12: note: expected ‘long unsigned int *’ but argument is of type ‘long long unsigned int *’
>  static int get_max_ploop_size(int version, unsigned int blocksize, __u64 *max)
>             ^
>

+++ lib/ploop.c
@@ -246,7 +246,7 @@ static int default_fmt_version(void)
 	return ploop_is_large_disk_supported() ? PLOOP_FMT_V2 : PLOOP_FMT_V1;
 }
 
-static int get_max_ploop_size(int version, unsigned int blocksize, __u64 *max)
+static int get_max_ploop_size(int version, unsigned int blocksize, unsigned long long *max)
 {
 	switch(version) {
 	case PLOOP_FMT_V1:
@@ -288,7 +288,7 @@ int ploop_get_max_size(unsigned int blocksize, unsigned long long *max)
 static int do_check_size(unsigned long long sectors, __u32 blocksize, int version,
 		__u64 max_fs_size)
 {
-	__u64 max;
+	unsigned long long max;
 
 	if (version == PLOOP_FMT_UNDEFINED)
 		return 0;
